# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardKpis {
  /** Total waste collected current month in metric tons */
  totalWasteCurrentMonth: number;
  /** Total waste collected last month in metric tons */
  totalWasteLastMonth: number;
  /** Percent change from last month */
  wasteTrend: number;
  /** Number of active collection operations today */
  activeOperationsToday: number;
  /** Landfill capacity used as percentage */
  capacityUsedPercent: number;
  /** Percentage of waste recycled or treated vs buried */
  diversionRate: number;
  /** Trend in diversion rate vs previous period */
  diversionRateTrend: number;
}

export interface TimeseriesPoint {
  /** ISO date string */
  date: string;
  /** Waste weight in metric tons */
  weightMt: number;
  /** @nullable */
  wasteType?: string | null;
}

export interface WasteCategoryBar {
  /** Waste category (Household, Industrial, Hazardous, Organic) */
  category: string;
  /** Collected volume in metric tons */
  collected: number;
  /** Successfully treated volume in metric tons */
  treated: number;
}

export interface TreatmentMethodSlice {
  /** Treatment method (Landfilled, Incinerated, Recycled, Composted) */
  method: string;
  /** Volume in metric tons */
  value: number;
  /** Percentage of total */
  percent: number;
}

export interface ForecastPoint {
  date: string;
  /** Cumulative landfill volume in metric tons */
  cumulativeVolume: number;
  isForecast?: boolean;
}

export interface ForecastResult {
  historicalPoints: ForecastPoint[];
  forecastPoints: ForecastPoint[];
  /** Maximum landfill capacity in metric tons */
  capacityMaxMt: number;
  /**
     * Months until capacity is reached, null if not within range
     * @nullable
     */
  warningMonths: number | null;
  /** Whether capacity will be exceeded within forecast window */
  willExceedCapacity: boolean;
}

export interface WasteLog {
  id: number;
  timestamp: string;
  weightMt: number;
  wasteType: string;
  wasteCode: string;
  truckId: string;
  site: string;
  treatmentStatus: string;
  treatmentMethod: string;
}

export interface WasteLogsPage {
  logs: WasteLog[];
  total: number;
  page: number;
  pageSize: number;
}

export interface AiInsightsRequest {
  /** JSON string of recent discharge statistics */
  statsJson: string;
}

export interface OperationalInsight {
  /** Category of insight (routing, recycling, anomaly, capacity) */
  category: string;
  title: string;
  description: string;
  /** info | warning | critical */
  severity: string;
}

export interface AiInsightsResult {
  insights: OperationalInsight[];
  generatedAt: string;
}

export interface FilterOptions {
  sites: string[];
  wasteTypes: string[];
}

export type GetDashboardKpisParams = {
dateFrom?: string;
dateTo?: string;
wasteType?: string;
site?: string;
};

export type GetDashboardTimeseriesParams = {
dateFrom?: string;
dateTo?: string;
wasteType?: string;
site?: string;
granularity?: GetDashboardTimeseriesGranularity;
};

export type GetDashboardTimeseriesGranularity = typeof GetDashboardTimeseriesGranularity[keyof typeof GetDashboardTimeseriesGranularity];


export const GetDashboardTimeseriesGranularity = {
  daily: 'daily',
  weekly: 'weekly',
} as const;

export type GetDashboardWasteCategoriesParams = {
dateFrom?: string;
dateTo?: string;
site?: string;
};

export type GetDashboardTreatmentMethodsParams = {
dateFrom?: string;
dateTo?: string;
wasteType?: string;
site?: string;
};

export type GetDashboardForecastParams = {
months?: number;
};

export type GetDashboardLogsParams = {
dateFrom?: string;
dateTo?: string;
wasteType?: string;
site?: string;
page?: number;
pageSize?: number;
};