# lib/api-zod/src/generated/types/aiInsightsResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { OperationalInsight } from './operationalInsight';

export interface AiInsightsResult {
  insights: OperationalInsight[];
  generatedAt: string;
}