# lib/api-zod/src/generated/types/dashboardKpis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardKpis {
  /** Total waste collected current month in metric tons */
  totalWasteCurrentMonth: number;
  /** Total waste collected last month in metric tons */
  totalWasteLastMonth: number;
  /** Percent change from last month */
  wasteTrend: number;
  /** Number of active collection operations today */
  activeOperationsToday: number;
  /** Landfill capacity used as percentage */
  capacityUsedPercent: number;
  /** Percentage of waste recycled or treated vs buried */
  diversionRate: number;
  /** Trend in diversion rate vs previous period */
  diversionRateTrend: number;
}