# lib/api-zod/src/generated/types/forecastPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface ForecastPoint {
  date: string;
  /** Cumulative landfill volume in metric tons */
  cumulativeVolume: number;
  isForecast?: boolean;
}