# lib/api-zod/src/generated/types/forecastResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { ForecastPoint } from './forecastPoint';

export interface ForecastResult {
  historicalPoints: ForecastPoint[];
  forecastPoints: ForecastPoint[];
  /** Maximum landfill capacity in metric tons */
  capacityMaxMt: number;
  /**
     * Months until capacity is reached, null if not within range
     * @nullable
     */
  warningMonths: number | null;
  /** Whether capacity will be exceeded within forecast window */
  willExceedCapacity: boolean;
}