# lib/api-zod/src/generated/types/operationalInsight.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface OperationalInsight {
  /** Category of insight (routing, recycling, anomaly, capacity) */
  category: string;
  title: string;
  description: string;
  /** info | warning | critical */
  severity: string;
}