# lib/api-zod/src/generated/types/timeseriesPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeseriesPoint {
  /** ISO date string */
  date: string;
  /** Waste weight in metric tons */
  weightMt: number;
  /** @nullable */
  wasteType?: string | null;
}