# lib/api-zod/src/generated/types/treatmentMethodSlice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface TreatmentMethodSlice {
  /** Treatment method (Landfilled, Incinerated, Recycled, Composted) */
  method: string;
  /** Volume in metric tons */
  value: number;
  /** Percentage of total */
  percent: number;
}