# lib/api-zod/src/generated/types/wasteLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface WasteLog {
  id: number;
  timestamp: string;
  weightMt: number;
  wasteType: string;
  wasteCode: string;
  truckId: string;
  site: string;
  treatmentStatus: string;
  treatmentMethod: string;
}