# lib/api-zod/src/generated/types/wasteLogsPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Landfill Discharges & Waste Operations Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { WasteLog } from './wasteLog';

export interface WasteLogsPage {
  logs: WasteLog[];
  total: number;
  page: number;
  pageSize: number;
}